Fix device number lookup bug (realloc could be confused)
e4rat-realloc sometimes skipped files with "not on an valid ext4 filesystem"
because it was confused by different devices having the same device numbers.
E.g. on Fedora 16-18 it might stumble upon /dev/watchdog0 instead of /dev/dm-0
because both have (major, minor) = (253, 0).  Similarly /dev/vcs2 and
/dev/loop2 both have (major, minor) = (7, 2).  In both cases, the difference is
that the former is a character device, while the latter is a block device.
This patch makes it look only for block devices, thus avoiding the collision.